javascript記述場所の変更
