+++ app/javascript/packs/application.js
@@ -6,6 +6,7 @@
 require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
+require("item")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
@@ -15,12 +16,4 @@ require("channels")
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 
-window.addEventListener('load', () => {
-  const priceInput = document.getElementById("item-price");
-  priceInput.addEventListener("input", () => {
-    const addTaxDom = document.getElementById("add-tax-price");
-    addTaxDom.innerHTML = Math.round(priceInput.value * 0.1 );
-    const addProfitDom = document.getElementById("profit");
-    addProfitDom.innerHTML = Math.round(priceInput.value - Math.round(priceInput.value * 0.1 ));
-  })
-});
+
